test: Testing updates work in progress
