--- packages/article-main-comment/__tests__/android/article-lead-asset.android.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-lead-asset.native";
-
-shared();
