+++ packages/article-main-comment/__tests__/ios/article-internal-components.ios.test.js
@@ -1,0 +1,3 @@
+import shared from "../shared-internal-components.native";
+
+shared();
