--- packages/article-main-comment/__tests__/ios/article-lead-asset.ios.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-lead-asset.native";
-
-shared();
