--- packages/article-main-comment/__tests__/ios/tracking.ios.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-tracking";
-
-shared();
