--- packages/article-main-comment/__tests__/shared-article-props.js
@@ -1,3 +1,0 @@
-export default {
-  refetch: () => {}
-};
