--- packages/article-main-comment/__tests__/shared-article-props.web.js
@@ -1,1 +1,0 @@
-export default {};
