+++ packages/article-main-comment/__tests__/web/article-internal-components.web.test.js
@@ -1,0 +1,3 @@
+import shared from "../shared-internal-components.web";
+
+shared();
