--- packages/article-main-comment/__tests__/web/article-lead-asset.web.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-lead-asset.web";
-
-shared();
