--- packages/article-main-comment/__tests__/web/tracking.web.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared-tracking";
-
-shared();
