+++ packages/article-main-comment/package.json
@@ -50,7 +50,6 @@
     "eslint": "4.19.1",
     "jest": "23.3.0",
     "jest-styled-components": "5.0.1",
-    "mockdate": "2.0.2",
     "prettier": "1.14.3",
     "react": "16.5.2",
     "react-dom": "16.5.2",
@@ -66,15 +65,11 @@
     "@times-components/article-error": "0.1.5",
     "@times-components/article-flag": "2.1.48",
     "@times-components/article-label": "2.2.20",
-    "@times-components/article-topics": "3.2.27",
-    "@times-components/caption": "2.1.46",
     "@times-components/context": "0.5.4",
     "@times-components/date-publication": "0.19.36",
     "@times-components/image": "4.3.3",
     "@times-components/styleguide": "3.7.2",
     "@times-components/utils": "4.0.10",
-    "@times-components/video": "4.1.11",
-    "@times-components/video-label": "2.1.48",
     "lodash.invert": "4.3.0",
     "prop-types": "15.6.2",
     "styled-components": "3.4.0"
